bingo_api/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pytesseract
Pillow
jinja2
pydantic